pyproject.toml
[project]
name = "hitl-pr-review"
version = "0.1.0"
description = "Human-in-the-Loop PR Review Agent with LangGraph"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.4.1",
    "langgraph-checkpoint-sqlite>=2.0.0",
    "langchain-openai>=0.3.0",
    "langchain-core>=0.3.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.28.0",
    "pydantic>=2.11.0",
    "python-dotenv>=1.1.0",
    "rich>=13.7.0",
    "typer>=0.15.0",
    "streamlit>=1.30.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["common", "audit"]
